+++ dl/tensor_t_test.go
@@ -8,7 +8,7 @@ import (
 func TestTensorTranspose(t *testing.T) {
 	tensor := &Tensor{
 		Shape: []int{2, 3},
-		Data:  []float64{1, 2, 3, 4, 5, 6},
+		Data:  []float32{1, 2, 3, 4, 5, 6},
 	}
 
 	transposed := tensor.Transpose([]int{1, 0})
